refactor: RulesetReplaceRoute
Changed RulesetEditRoute to RulesetReplaceRoute
