--- src/routes/RulesetEditRoute/index.js
@@ -1,1 +1,0 @@
-export { default } from './RulesetEditRoute';
