+++ src/routes/index.js
@@ -2,6 +2,6 @@ export { default as SerialsRoute } from './SerialsRoute';
 export { default as SerialCreateRoute } from './SerialCreateRoute';
 export { default as RulesetViewRoute } from './RulesetViewRoute';
 export { default as RulesetCreateRoute } from './RulesetCreateRoute';
-export { default as RulesetEditRoute } from './RulesetEditRoute';
+export { default as RulesetReplaceRoute } from './RulesetReplaceRoute';
 export { default as SerialEditRoute } from './SerialEditRoute';
 export { default as PieceSetsRoute } from './PieceSetsRoute';
